requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dataclasses-json==0.6.7
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langsmith==0.2.11
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.14
outcome==1.3.0.post0
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.10.1
PySocks==1.7.1
python-crontab==3.2.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.27.1
sentence-transformers==3.3.1
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typing-inspect==0.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.51
